smarty_sloths/requirements.txt
Django==1.10.4
Pillow==3.4.2
pytesseract==0.1.6
django-extensions==1.7.5
pyOpenSSL==16.2.0
Werkzeug==0.11.11
